+++ R/attach_metadata.R
@@ -5,7 +5,7 @@
 #' 
 #' @aliases attach_metadata
 #' @param peak_table A `peak_table` object.
-#' @param metadata A `data.frame` containing the sample meta-data.
+#' @param metadata A `data.frame` containing the sample metadata.
 #' @param column The name of the column containing the sample names.
 #' @return A \code{peak_table} object with attached metadata in the \code{
 #' $sample_meta} slot.
@@ -19,19 +19,20 @@
 #' @export attach_metadata
 
 attach_metadata <- function(peak_table, metadata, column){
+  check_peaktable(peak_table)
   if (any(grepl("tbl", class(metadata)))){
     metadata <- as.data.frame(metadata)
   }
   if (!inherits(metadata, "data.frame")){
-    stop("Please provide metadata as a data.frame")
+    stop("Please provide metadata as a `data.frame`")
   }
   if (!(column %in% colnames(metadata)))
     stop(paste0("Column, ", column, ", is not found."))
   if (sum((duplicated(metadata[,column]))) > 0)
-    stop(paste0("Sample names must be unique. Please check column '", column,
-    "' for duplicates."))
+    stop(paste("Sample names must be unique. Please check column", sQuote(column),
+    "for duplicates."))
   if (!inherits(peak_table,"peak_table"))
-    stop(paste("Provided peak_table object must be of class 'peak_table'."))
+    stop(paste("Provided peak table object must be of class 'peak_table'."))
   meta <- data.frame(rownames(peak_table$tab))
   names(meta) <- column
   metadata[, column] <- as.character(metadata[, column])
@@ -44,9 +45,9 @@ attach_metadata <- function(peak_table, metadata, column){
   return(peak_table)
 }
 
-
 #' note: convenience function from stackoverflow:
 #' https://stackoverflow.com/questions/17878048/merge-two-data-frames-while-keeping-the-original-row-order
+#' @noRd
 keep_order <- function(data, fn, ...) { 
   col <- ".sortColumn"
   data[,col] <- 1:nrow(data) 
@@ -58,7 +59,7 @@ keep_order <- function(data, fn, ...) {
 } 
 
 
-#' Gather reference spectra.
+#' Get reference spectra.
 #' 
 #' Defines reference spectra. Reference spectra are defined either as the 
 #' spectrum with the highest intensity (\code{max.int}) or as the spectrum 
@@ -78,16 +79,17 @@ keep_order <- function(data, fn, ...) {
 #' @seealso \code{\link{get_peaks}}
 #' @noRd
 
-gather_reference_spectra <- function(peak_table, chrom_list,
+get_reference_spectra <- function(peak_table, chrom_list,
                                      ref = c("max.cor", "max.int")){
+  check_peaktable(peak_table)
   if (!inherits(peak_table, "peak_table"))
-    stop(paste("Provided peak_table object must be of class 'peak_table'."))
+    stop("Provided peak_table object must be a `peak_table` object.")
   if (missing(chrom_list)){
     chrom_list <- get_chrom_list(peak_table)
   } else get_chrom_list(peak_table, chrom_list)
   ref <- match.arg(ref, c("max.cor", "max.int"))
-  X<-colnames(peak_table$tab)
-  if (ref=="max.cor"){
+  X <- colnames(peak_table$tab)
+  if (ref == "max.cor"){
     sp.l <- lapply(X,function(pk){
       plot_all_spectra(peak = pk, peak_table, chrom_list,
                        plot_spectrum = FALSE, export_spectrum = TRUE,
@@ -100,7 +102,8 @@ gather_reference_spectra <- function(peak_table, chrom_list,
       plot_spectrum(loc = pk, peak_table, chrom_list, plot_trace=FALSE,
                     plot_spectrum = FALSE, export_spectrum = TRUE,
                     verbose = FALSE,
-                    scale_spectrum = TRUE)})
+                    scale_spectrum = TRUE, engine = "base")
+      })
     sp.ref <- do.call(cbind, sp.ref)
   }
   colnames(sp.ref) <- colnames(peak_table$tab)
@@ -131,16 +134,18 @@ gather_reference_spectra <- function(peak_table, chrom_list,
 #' pk_tab <- attach_ref_spectra(pk_tab, ref="max.int")
 #' pk_tab <- attach_ref_spectra(pk_tab, ref = "max.cor")
 #' @export attach_ref_spectra
-#' 
+
 attach_ref_spectra <- function(peak_table, chrom_list, ref = c("max.cor","max.int")){
-  peak_table$ref_spectra <- gather_reference_spectra(peak_table, chrom_list, ref)
+  check_peaktable(peak_table)
+  peak_table$ref_spectra <- get_reference_spectra(peak_table, chrom_list, ref)
+  peak_table$args["reference_spectra"] <- ref
   return(peak_table)
 }
 
 #' Normalize peak table or chromatograms
 #' 
 #' Normalizes peak table or list of chromatograms by specified column in sample
-#' meta-data. Metadata must first be attached to \code{peak_table} using
+#' metadata. Metadata must first be attached to \code{peak_table} using
 #' \code{\link{attach_metadata}}.
 #' 
 #' @param peak_table A `peak_table` object
@@ -148,6 +153,8 @@ attach_ref_spectra <- function(peak_table, chrom_list, ref = c("max.cor","max.in
 #' @param chrom_list List of chromatograms for normalization. The samples must
 #' be in same order as the peak_table.
 #' @param what `peak_table` or list of chromatograms (`chrom_list`).
+#' @param by Whether to normalize by a column in sample metadata (\code{meta}) or
+#' by a column in the peak table itself (\code{peak}).
 #' @return A \code{peak_table} object where the peaks are normalized by the mass
 #' of each sample.
 #' @author Ethan Bass
@@ -161,20 +168,23 @@ attach_ref_spectra <- function(peak_table, chrom_list, ref = c("max.cor","max.in
 #' @export normalize_data
 
 normalize_data <- function(peak_table, column, chrom_list,
-                           what=c('peak_table','chrom_list')){
-  if (!inherits(peak_table,"peak_table"))
-    stop(paste("Provided peak_table object must be of class 'peak_table'."))
+                           what = c('peak_table','chrom_list'),
+                           by=c("meta", "peak")){
+  check_peaktable(peak_table)
   if (!is.data.frame(peak_table$sample_meta))
-    stop("Meta-data must be attached to peak_table prior to normalization.")
+    stop("Metadata must be attached to peak_table prior to normalization.")
   if (!(column %in% colnames(peak_table$sample_meta)))
-    stop(paste0("Column, ", column, ", is not found."))
+    stop(paste0("The specified column (", sQuote(column), ") could not be found."))
   what <- match.arg(what, c("peak_table", "chrom_list"))
+  by <- match.arg(by, c("meta", "peak"))
+  df <- switch(by, meta = peak_table$sample_meta, peak = peak_table$tab)
   if (what == "peak_table"){
     pktab <- as.data.frame(t(sapply(seq_len(nrow(peak_table$tab)), function(samp){
-      as.numeric(as.vector(peak_table$tab[samp,]))/peak_table$sample_meta[samp,column]
+      as.numeric(as.vector(peak_table$tab[samp,]))/df[samp,column]
     })))
     rownames(pktab) <- rownames(peak_table$tab)
     peak_table$tab <- pktab
+    peak_table$args[c("normalized","normalization_by")] <- c(TRUE, column)
     return(peak_table)
   } else if (what == "chrom_list"){
     if (missing(chrom_list)){
@@ -183,7 +193,7 @@ normalize_data <- function(peak_table, column, chrom_list,
     if (mean(elementwise.all.equal(names(chrom_list), rownames(peak_table$tab))) < 1)
       stop("Names of chromatograms do not match the peak table.")
     chrom_list <- lapply(seq_len(nrow(peak_table$tab)), function(samp){
-      chrom_list[[samp]]/peak_table$sample_meta[samp,column]
+      chrom_list[[samp]]/df[samp,column]
     })
     return(chrom_list)
   }